Custom.nsh
All modifications to the PortableApps.com Launcher which do not involve debugging go in a file Custom.nsh in the Other\Source directory of a package.

Changed in version 2.1: previously this file was Other\Source\PortableApps.comLauncherCustom.nsh. The Generator will move this file to the new location if it exists.

http://portableapps.com/manuals/PortableApps.comLauncher/advanced/custom.html